+++ nacc_form_validator/json_logic.py
@@ -63,13 +63,13 @@ def less(a, b, *args):
     # Handling empty values
     if a is None and b is None:
         return False
-    
+
     if a is None:
         return True
-    
+
     if b is None:
         return False
-    
+
     types = {type(a), type(b)}
     if float in types or int in types:
         try: