+++ CHANGELOG.md
@@ -12,7 +12,7 @@ Documentation of release versions of `nacc-form-validator`
 * Adds `previous_record` as a special keyword for `__get_value_for_key`
 * Adds `get_previous_record` method to grab previous record from Datastore
 * Updates `_check_subschema_valid` to accept an optional record parameter, defaults to the document - used for rules that require the previous record
-* Upddates `compare_with` to add `ignore_blank` option - if True, does not return a validation error if the base is None. Defaults to False
+* Updates `compare_with` to add `ignore_blank` option - if True, does not return a validation error if the base is None. Defaults to False
 
 ## 0.2.0