+++ CHANGELOG.md
@@ -2,12 +2,14 @@
 
 Documentation of release versions of `nacc-form-validator`
 
-## 0.3.1
+## 0.4.0
 
 * Updates `_check_subschema_valid` to accept an optional record parameter, defaults to the document - used for rules that require the previous record
 * Updates `compare_with` to support the `abs` operator
+* Moves logic for `compare_values` to its own utility method
 * Adds `previous_record` as a special keyword for `compare_with`
 * Adds `get_previous_record` method to grab previous record from Datastore, which can grab the previous record or the previous record where a specific field is non-empty
+* Adds new rule `compare_dates` to handle rules that need to compare dates
 
 ## 0.3.0